Disable get_proxy_ip() SSL check and move to utils
get_proxy_ip() SSL check is now disabled due to it causing problems on
some cases. The function is also moved to utils.py since it has no
special meaning in putty_controllers.py
